working_draft/conformance/buffimgs/WEBCL_WebCLCommandQueue_enqueueWriteBufferRect_004.html

<!--
 Copyright (C) 2012 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCL_High</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
/******WebCLCommandQueue enqueueWriteBufferRect with invalid WebCLBuffer Obj(, bufferOrigin(2D), hostOrigin(2D), region(2D), bufferRowPitch, bufferSlicePitch, hostRowPitch and arrBuffer*************/
try {
	debug("WebCLCommandQueue enqueueWriteBufferRect with invalid WebCLBuffer Obj, bufferOrigin(2D), hostOrigin(2D), region(2D), bufferRowPitch, bufferSlicePitch, hostRowPitch and arrBuffer");

	var context = webcl.createContext();
	if( context instanceof WebCLContext ) {
		try {
			var webCLCommandQueue = context.createCommandQueue();
			if(webCLCommandQueue instanceof WebCLCommandQueue) {
				var bufferOrigin = new Int32Array(3);
				bufferOrigin[0] = 12;
				bufferOrigin[1] = 25;
				bufferOrigin[2] = 0;
				var hostOrigin = new Int32Array(3);
				hostOrigin[0] = 12;
				hostOrigin[1] = 25;
				hostOrigin[2] = 0;
				var region = new Int32Array(3);
				region[0] = 10;
				region[1] = 20;
				region[2] = 1;
				var arrBuffer = new ArrayBuffer(32); // 32-byte buffer
				try {
					webCLCommandQueue.enqueueWriteBufferRect(invalid_object,
															true,
															bufferOrigin,
															hostOrigin,
															region,
															Float32Array.BYTES_PER_ELEMENT * 320 * 4,
															Float32Array.BYTES_PER_ELEMENT * 320 * 4,
															Float32Array.BYTES_PER_ELEMENT * 320 * 4,
															Float32Array.BYTES_PER_ELEMENT * 320 * 4,
															arrBuffer);
					testFailed("It should throw an exception.");notifyFinishedToHarness();
				} catch(e) {
					shouldBeEqual(e.name, "INVALID_MEM_OBJECT");
					notifyFinishedToHarness();
				}
			} else {
				testFailed("createCommandQueue is returned : "+webCLCommandQueue);
				notifyFinishedToHarness();
			}
		} catch(e) {
			testFailed("context createCommandQueue() throws an exception : "+e.message);
			notifyFinishedToHarness();
		}
	} else {
		testFailed("createContext returned : "+context);
		notifyFinishedToHarness();
	}
} catch (e) {
	testFailed("createContext throws an exception : "+e.message);
	notifyFinishedToHarness();
}
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
